hydride_segmentation_tool/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hydride-segmentation-tool"
version = "0.1.0"
description = "Offline hydride segmentation tool"
authors = [{ name = "Hydride Team" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Pillow",
    "numpy",
    "opencv-python",
    "pytest",
    "tk"
]

[project.optional-dependencies]
dev = ["pytest"]

